Move UpcastingDistinctList into a separate file
Rename to ProxyDistinctList
